9. AI_SEO_Content_Optimizer/requirements.txt
requests==2.31.0
beautifulsoup4>=4.12.2,<4.13.0
openai==1.12.0
python-dotenv>=1.0.0,<2.0.2
tqdm==4.66.1
colorama==0.4.6
rich>=13.7.0,<13.8.0 